Make queryPlaylistsByName take InclusionOptions
Allows it to account for including playlists or not. This is mostly
unnecessary vs. just removing the library filter altogether as when FIL
would go to import an animations playlist without "Include Animations"
selected no titles belonging to that playlist would have been imported
so it would be empty and therefore not saved.

Regardless this is being done for posterity/cleanliness, just know it
can be removed if required.