Don't append .exe to server/start-stop entries
This was originally done because it was believed that QProcess required
a full filename in order to be able to start a process correctly, but
testing on Windows shows that just the base name of the executable will
work.

This can't be done on Linux outright given the '.exe' extension is not
used. If for some reason this proves to be a problem on windows then it
must be done conditionally based on platform, but for now it is being
removed outright since the above testing indicates it shouldn't ever be
necessary.